First Update
This is the first batch of work I have done with Open CV.
Currently these are the two files that work:

detect_demo.py : this is a program that parses through every foot in
the Pictures directory, recognizes where the faces are, draws
rectangles around the faces, and writes each new image to the
OutputPictures directory. There are some sample photos that I have ran
in the directories provided.

video_detect.py: This is a program that captures a live video feed and
detects faces in the live video feed, and draws rectangles around them
in real time.

NOTE: You must have OpenCV, Numpy, and Scipy installed in your Python
pathway in order for these programs to function.